requirements.txt
altair==4.2.0
numpy
pandas
pydeck
google-cloud-firestore
streamlit==1.37.0
